register/users/account component css has been done

+++ src/app/components/users/users.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, AfterViewInit } from "@angular/core";
+import { Component, OnInit, ElementRef, AfterViewInit } from "@angular/core";
 import { DataService } from "../../services/data.service";
 import { Route, Router, Data } from "@angular/router";
 import { HttpErrorResponse } from "@angular/common/http";
@@ -15,36 +15,63 @@ export class UsersComponent implements OnInit {
   profile;
   userType;
 
-  constructor(private router: Router, private dataservice: DataService) {}
+  constructor(
+    private router: Router,
+    private dataservice: DataService,
+    private elementRef: ElementRef
+  ) {}
 
   ngOnInit() {
+    this.elementRef.nativeElement.ownerDocument.body.style.backgroundColor =
+      "#fff";
+
     this.profile = JSON.parse(sessionStorage.getItem("user"));
     this.userType = false;
-    if (this.profile.type == "super") {
-      this.userType = true;
-    }
+    if (this.profile == undefined) {
+      this.router.navigateByUrl("/login");
+    } else {
+      if (this.profile.type == "super") {
+        this.userType = true;
+      }
 
-    if (this.profile.type == "normal") {
-      alert("Only for Super/Group Admin");
-      this.router.navigateByUrl("/account");
-    } else if (this.profile.type == "group assis") {
-      alert("Only for Super/Group Admin");
-      this.router.navigateByUrl("/account");
-    }
+      if (this.profile.type == "normal") {
+        alert("Only for Super/Group Admin");
+        this.router.navigateByUrl("/account");
+      } else if (this.profile.type == "group assis") {
+        alert("Only for Super/Group Admin");
+        this.router.navigateByUrl("/account");
+      }
 
-    this.dataservice.getUsers().subscribe(data => {
-      this.users = data;
-    }),
-      (error: HttpErrorResponse) => {
-        alert("Error");
-      };
+      this.dataservice.getUsers().subscribe(data => {
+        this.users = data;
+      }),
+        (error: HttpErrorResponse) => {
+          alert("Error");
+        };
+    }
   }
   ngAfterViewInit() {}
 
   //delete user if clicked
-  deleteUser(email: string) {
-    this.dataservice.deleteUser(email).subscribe(data => {
-      this.users = data;
-    });
+  deleteUser(email: string, type: string) {
+    console.log(type);
+    if (type !== "super") {
+      this.dataservice.deleteUser(email).subscribe(data => {
+        this.users = data;
+      });
+    } else {
+      alert("Dont Delelte Me");
+    }
+  }
+
+  //give super to user
+  giveSuper(user) {
+    if (user.type == "super") {
+      alert("already type super");
+    } else {
+      this.dataservice.giveSuper(user).subscribe(data => {
+        this.users = data;
+      });
+    }
   }
 }
